Update to new UI
IMDB changed their UI to a mobile-only variant awhile ago, the
changes to the DOM broke the plugin. This commit updates to use
the new one. This branch will replace master as the default branch
going forward unless IMDB was to revert back to the old UI.

Also update to manifest v3.